+++ DotNet/DataAcquisition/Jobs/AcquisitionProcessingJob.cs
@@ -4,7 +4,6 @@ using LantanaGroup.Link.DataAcquisition.Domain.Application.Models;
 using LantanaGroup.Link.DataAcquisition.Domain.Application.Models.Kafka;
 using LantanaGroup.Link.DataAcquisition.Domain.Application.Queries;
 using LantanaGroup.Link.DataAcquisition.Domain.Settings;
-using LantanaGroup.Link.Shared.Application.Error.Interfaces;
 using LantanaGroup.Link.Shared.Application.Models;
 using LantanaGroup.Link.Shared.Application.Services.Security;
 using Quartz;
@@ -20,20 +19,17 @@ public class AcquisitionProcessingJob : IJob
     private readonly ILogger<AcquisitionProcessingJob> _logger;
     private readonly IServiceScopeFactory _serviceScopeFactory;
     private readonly IProducer<string, ReadyToAcquire> _readyToAcquireProducer;
-    protected readonly ITransientExceptionHandler<string, ReadyToAcquire> _transientExceptionHandler;
     private readonly IProducer<string, ResourceAcquired> _resourceAcquiredProducer;
 
     public AcquisitionProcessingJob(
         ILogger<AcquisitionProcessingJob> logger,
         IServiceScopeFactory serviceScopeFactory,
         IProducer<string, ReadyToAcquire> readyToAcquireProducer,
-        ITransientExceptionHandler<string, ReadyToAcquire> transientExceptionHandler,
         IProducer<string, ResourceAcquired> resourceAcquiredProducer)
     {
         _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         _serviceScopeFactory = serviceScopeFactory ?? throw new ArgumentNullException(nameof(serviceScopeFactory));
         _readyToAcquireProducer = readyToAcquireProducer ?? throw new ArgumentNullException(nameof(readyToAcquireProducer));
-        _transientExceptionHandler = transientExceptionHandler ?? throw new ArgumentNullException(nameof(transientExceptionHandler));
         _resourceAcquiredProducer = resourceAcquiredProducer ?? throw new ArgumentNullException(nameof(resourceAcquiredProducer));
     }
 
@@ -43,7 +39,7 @@ public class AcquisitionProcessingJob : IJob
         await ProcessPendingTailingMessages(context.CancellationToken);
     }
 
-    private async Task ProcessPendingLogs(CancellationToken cancellationToken)
+    public async Task ProcessPendingLogs(CancellationToken cancellationToken)
     {
         try
         {
@@ -61,7 +57,6 @@ public class AcquisitionProcessingJob : IJob
 
     private async Task ProcessFacilityPendingLogs(string facilityId, CancellationToken cancellationToken)
     {
-        ReadyToAcquire messageValue = null;
         try
         {
             using var scope = _serviceScopeFactory.CreateScope();
@@ -71,27 +66,26 @@ public class AcquisitionProcessingJob : IJob
 
             var config = await fhirQueryConfigurationManager.GetAsync(facilityId, cancellationToken);
 
-            const int pageSize = 100;
-            var baseMessage = "Request FAILED due to missing FhirQueryConfiguration. FacilityId:";
+            const int pageSize = 25;
             if (config == null)
             {
-                _logger.LogCritical("{baseMessage} {facilityId}", baseMessage.Sanitize(), facilityId.Sanitize());
+                _logger.LogCritical("Request FAILED due to missing FhirQueryConfiguration. FacilityId: {facilityId}", facilityId.Sanitize());
 
-                int page = 1;
+                string? lastMissingConfigId = null;
                 while (true)
                 {
-                    var requests = await dataAcquisitionLogQueries.GetPendingAndRetryableFailedRequestsForFacility(facilityId, page, pageSize, cancellationToken);
+                    var requests = await dataAcquisitionLogQueries.GetNextEligibleBatchForFacility(facilityId, lastMissingConfigId, pageSize, cancellationToken);
                     if (!requests.Any()) break;
 
                     foreach (var request in requests)
                     {
                         request.Status = RequestStatus.Failed;
                         request.Notes ??= new List<string>();
-                        request.Notes.Add($"{baseMessage} {request.FacilityId}.");
+                        request.Notes.Add($"[{DateTime.UtcNow}] Request FAILED due to missing FhirQueryConfiguration. FacilityId: {request.FacilityId}.");
                         await dataAcquisitionLogManager.UpdateAsync(request, cancellationToken);
                     }
 
-                    page++;
+                    lastMissingConfigId = requests.Last().Id;
                 }
 
                 return;
@@ -103,11 +97,16 @@ public class AcquisitionProcessingJob : IJob
                 return;
             }
 
-            int pageNumber = 1;
+            string? lastId = null;
             while (true)
             {
-                var requests = await dataAcquisitionLogQueries.GetPendingAndRetryableFailedRequestsForFacility(facilityId, pageNumber, pageSize, cancellationToken);
-                if (!requests.Any()) break;
+                _logger.LogInformation("Fetching batch after Id {lastId} for facility {facilityId}", lastId ?? "null", facilityId);
+                var requests = await dataAcquisitionLogQueries.GetNextEligibleBatchForFacility(facilityId, lastId, pageSize, cancellationToken);
+                if (!requests.Any())
+                {
+                    _logger.LogInformation("No more logs to process for facility {facilityId}", facilityId);
+                    break;
+                }
 
                 _logger.BeginScope("Processing {count} processable requests for facility {facilityId}", requests.Count, facilityId);
 
@@ -130,9 +129,9 @@ public class AcquisitionProcessingJob : IJob
                         request.Notes.Add($"[{DateTime.UtcNow}] Retrying failed request. Attempt {request.RetryAttempts}.");
                     }
 
-                    messageValue = new ReadyToAcquire { FacilityId = facilityId, LogId = request.Id };
+                    var messageValue = new ReadyToAcquire { FacilityId = facilityId, LogId = request.Id };
 
-                    _logger.LogInformation("Generating ReadyToAcquire message for log id: {request.Id}", request.Id.Sanitize());
+                    _logger.LogInformation("Generating ReadyToAcquire message for log id: {requestId}", request.Id.Sanitize());
 
                     request.Status = RequestStatus.Ready;
                     await dataAcquisitionLogManager.UpdateAsync(request, cancellationToken);
@@ -142,9 +141,9 @@ public class AcquisitionProcessingJob : IJob
                         _logger.LogInformation("Producing ReadyToAcquire message for log id: {logId} and facility id: {facilityId}", request.Id.Sanitize(), facilityId.Sanitize());
 
                         var headers = new Headers
-                        {
-                            { "X-Correlation-Id", Encoding.UTF8.GetBytes(request.CorrelationId?.ToString() ?? string.Empty) }
-                        };
+                    {
+                        { "X-Correlation-Id", Encoding.UTF8.GetBytes(request.CorrelationId?.ToString() ?? string.Empty) }
+                    };
 
                         await _readyToAcquireProducer.ProduceAsync(
                             KafkaTopic.ReadyToAcquire.ToString(),
@@ -169,11 +168,9 @@ public class AcquisitionProcessingJob : IJob
                         request.Notes.Add($"[{DateTime.UtcNow}] Failed to produce ReadyToAcquire message: {ex.Message}");
                         await dataAcquisitionLogManager.UpdateAsync(request, cancellationToken);
                     }
-
-                    messageValue = null;
                 }
 
-                pageNumber++;
+                lastId = requests.Last().Id;
             }
 
             _logger.LogInformation("Completed processing processable requests for facility {facilityId}.", facilityId);
@@ -181,7 +178,6 @@ public class AcquisitionProcessingJob : IJob
         catch (Exception ex)
         {
             _logger.LogError(ex, "Error processing acquisition job for facility id: {facilityId}", facilityId);
-            _transientExceptionHandler.HandleException(ex, messageValue, facilityId, $"Error processing acquisition job for facility id: {facilityId}");
         }
     }
 
@@ -205,7 +201,7 @@ public class AcquisitionProcessingJob : IJob
         return currentTime >= minAcquisitionPullTime || currentTime <= maxAcquisitionPullTime;
     }
 
-    private async Task ProcessPendingTailingMessages(CancellationToken cancellationToken)
+    public async Task ProcessPendingTailingMessages(CancellationToken cancellationToken)
     {
         using var scope = _serviceScopeFactory.CreateScope();
         var dataAcquisitionLogManager = scope.ServiceProvider.GetRequiredService<IDataAcquisitionLogManager>();
