--- DotNet/DataAcquisitionTests/Resources/AllergyIntolerance.txt
@@ -1,20 +1,0 @@
-{
-  "resourceType": "AllergyIntolerance",
-  "id": "allergy-1",
-  "clinicalStatus": {
-    "coding": [
-      {
-        "system": "http://terminology.hl7.org/CodeSystem/allergyintolerance-clinical",
-        "code": "active"
-      }
-    ]
-  },
-  "patient": {
-    "reference": "Patient/patient-1"
-  },
-  "reaction": [
-    {
-      "description": "Initial Description"
-    }
-  ]
-}
