--- DotNet/DataAcquisitionTests/Resources/BodyTempObservation.txt
@@ -1,18 +1,0 @@
-{
-  "resourceType": "Observation",
-  "id": "obs-123",
-  "status": "final",
-  "code": {
-    "coding": [
-      {
-        "system": "http://loinc.org",
-        "code": "8310-5",
-        "display": "Body temperature"
-      }
-    ]
-  },
-  "subject": {
-    "reference": "Patient/123"
-  },
-  "effectiveDateTime": "2025-05-19T12:00:00Z"
-}
