--- DotNet/DataAcquisitionTests/Resources/Condition.txt
@@ -1,19 +1,0 @@
-{
-  "resourceType": "Condition",
-  "id": "condition-1",
-  "clinicalStatus": {
-    "coding": [
-      {
-        "system": "http://terminology.hl7.org/CodeSystem/condition-clinical",
-        "code": "active"
-      }
-    ]
-  },
-  "code": {
-    "text": "Placeholder"
-  },
-  "subject": {
-    "reference": "Patient/patient-1"
-  },
-  "onsetDateTime": "2023-05-01T10:00:00Z"
-}
