--- DotNet/DataAcquisitionTests/Resources/DiabetesCondition.txt
@@ -1,25 +1,0 @@
-{
-  "resourceType": "Condition",
-  "id": "cond-456",
-  "clinicalStatus": {
-    "coding": [
-      {
-        "system": "http://terminology.hl7.org/CodeSystem/condition-clinical",
-        "code": "active"
-      }
-    ]
-  },
-  "code": {
-    "coding": [
-      {
-        "system": "http://snomed.info/sct",
-        "code": "44054006",
-        "display": "Diabetes mellitus type 2"
-      }
-    ]
-  },
-  "subject": {
-    "reference": "Patient/123"
-  },
-  "onsetDateTime": "2025-01-01"
-}
