--- DotNet/DataAcquisitionTests/Resources/DiagnosticReport.txt
@@ -1,12 +1,0 @@
-{
-  "resourceType": "DiagnosticReport",
-  "id": "diagreport-1",
-  "status": "final",
-  "code": {
-    "text": "Initial Code"
-  },
-  "subject": {
-    "reference": "Patient/patient-1"
-  },
-  "effectiveDateTime": "2023-07-01T09:00:00Z"
-}
