--- DotNet/DataAcquisitionTests/Resources/EvaluatedEncounter1.txt
@@ -1,45 +1,0 @@
-﻿{
-	"resourceType": "Encounter",
-	"id": "Encounter-multi2",
-	"meta": {
-	    "versionId": "10",
-	    "lastUpdated": "2024-12-10T16:21:26.443+00:00",
-	    "source": "#f5dd83a50f95f19c",
-	    "profile": [
-	        "http://hl7.org/fhir/us/core/StructureDefinition/us-core-encounter"
-	    ]
-	},
-	"status": "arrived",
-	"class": {
-	    "system": "http://terminology.hl7.org/CodeSystem/v3-ActCode",
-	    "code": "ACUTE",
-	    "display": "outpatient acute"
-	},
-	"type": [
-	    {
-	        "coding": [
-	            {
-	                "system": "http://snomed.info/sct",
-	                "code": "5"
-	            }
-	        ],
-	        "text": "Emergency hospital admission (procedure)"
-	    }
-	],
-	"period": {
-	    "start": "2024-12-01T10:00:00.000Z",
-	    "end": "2024-12-31T10:00:00.000Z"
-	},
-	"location": [
-	    {
-	        "location": {
-	            "reference": "Location/Location-multi1"
-	        },
-	        "status": "active",
-	        "period": {
-	            "start": "2024-12-01T10:00:00.000Z",
-	            "end": "2024-12-31T10:00:00.000Z"
-	        }
-	    }
-	]
-}
