--- DotNet/DataAcquisitionTests/Resources/MedicationRequest.txt
@@ -1,25 +1,0 @@
-{
-  "resourceType": "MedicationRequest",
-  "id": "medreq-1",
-  "status": "active",
-  "intent": "order",
-  "medicationCodeableConcept": {
-    "text": "Aspirin"
-  },
-  "subject": {
-    "reference": "Patient/patient-1"
-  },
-  "dosageInstruction": [
-    {
-      "doseAndRate": [
-        {
-          "doseQuantity": {
-            "value": 325,
-            "unit": "mg"
-          }
-        }
-      ]
-    }
-  ],
-  "note": []
-}
