--- DotNet/DataAcquisitionTests/Resources/Observation.txt
@@ -1,11 +1,0 @@
-{
-  "resourceType": "Observation",
-  "id": "obs-1",
-  "status": "final",
-  "code": { "text": "Blood Pressure" },
-  "valueQuantity": { "value": 120, "unit": "mmHg" },
-  "component": [
-    { "code": { "text": "Systolic" }, "valueQuantity": { "value": 0, "unit": "mmHg" } },
-    { "code": { "text": "Diastolic" }, "valueQuantity": { "value": 0, "unit": "mmHg" } }
-  ]
-}
