--- DotNet/DataAcquisitionTests/Resources/PatientBirthDateTest.txt
@@ -1,10 +1,0 @@
-{
-  "resourceType": "Patient",
-  "id": "patient-1",
-  "name": [
-    {
-      "text": "Initial Name"
-    }
-  ],
-  "birthDate": "1990-01-01"
-}
