--- DotNet/DataAcquisitionTests/Resources/PatientJohnDoe.txt
@@ -1,11 +1,0 @@
-{
-  "resourceType": "Patient",
-  "id": "patient-1",
-  "name": [
-    {
-      "given": ["John"],
-      "family": "Doe"
-    }
-  ],
-  "extension": []
-}
