+++ link-cloud.sln
@@ -50,8 +50,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceTests", "DotNet\Serv
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataAcquisition.AcquisitionWorker", "DotNet\DataAcquisition.AcquisitionWorker\DataAcquisition.AcquisitionWorker.csproj", "{FDCF876C-112F-48E4-A33F-13A4C1AED4BF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataAcquisitionTests", "DotNet\DataAcquisitionTests\DataAcquisitionTests.csproj", "{2C221433-A8A0-3414-58DC-701892FAB3C2}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BackendE2ETests", "Tests\BackendE2ETests\BackendE2ETests.csproj", "{E34B41C5-3093-05C2-C945-BCB49AC34E7F}"
 EndProject
 Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-compose.dcproj", "{0FE36D9C-71D5-440F-B676-2A158E9D596D}"
@@ -168,12 +166,6 @@ Global
 		{FDCF876C-112F-48E4-A33F-13A4C1AED4BF}.NoVariablesDefault|Any CPU.Build.0 = Release|Any CPU
 		{FDCF876C-112F-48E4-A33F-13A4C1AED4BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FDCF876C-112F-48E4-A33F-13A4C1AED4BF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2C221433-A8A0-3414-58DC-701892FAB3C2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2C221433-A8A0-3414-58DC-701892FAB3C2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2C221433-A8A0-3414-58DC-701892FAB3C2}.NoVariablesDefault|Any CPU.ActiveCfg = Release|Any CPU
-		{2C221433-A8A0-3414-58DC-701892FAB3C2}.NoVariablesDefault|Any CPU.Build.0 = Release|Any CPU
-		{2C221433-A8A0-3414-58DC-701892FAB3C2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2C221433-A8A0-3414-58DC-701892FAB3C2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E34B41C5-3093-05C2-C945-BCB49AC34E7F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E34B41C5-3093-05C2-C945-BCB49AC34E7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E34B41C5-3093-05C2-C945-BCB49AC34E7F}.NoVariablesDefault|Any CPU.ActiveCfg = Release|Any CPU
@@ -195,7 +187,6 @@ Global
 		{C4070261-D248-4CA5-A532-6AAA71AACC35} = {37C302AB-8C98-4CC8-828E-0647AF098586}
 		{472A561B-EF7F-467E-96ED-1FB130E43292} = {C98B4989-1078-4632-BC7D-9CE52A40B360}
 		{F25C619A-50DC-4A67-9406-F8A2DB07AA22} = {554985B2-5CA6-49D3-8F07-08080C4EACEA}
-		{2C221433-A8A0-3414-58DC-701892FAB3C2} = {554985B2-5CA6-49D3-8F07-08080C4EACEA}
 		{E34B41C5-3093-05C2-C945-BCB49AC34E7F} = {554985B2-5CA6-49D3-8F07-08080C4EACEA}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
